+++ android/app/src/main/java/com/funnyhatsoftware/spacedock/data/DataLoader.java
@@ -163,10 +163,10 @@ public class DataLoader extends DefaultHandler {
 
             @Override
             public void put(String externalId, SetItem s) {
-            	universe.admirals.put(externalId, (Admiral) s);
+                universe.admirals.put(externalId, (Admiral) s);
             }
 
-			@Override
+            @Override
             public void afterUpdate(SetItem s) {
             }