+++ ios/DockOfficerRowHandler.h
@@ -1,0 +1,6 @@
+#import "DockRowHandler.h"
+
+@class DockEquippedUpgrade;
+
+@interface DockOfficerRowHandler : DockRowHandler
+@end