+++ src/DockUpgrade+Addons.h
@@ -16,6 +16,7 @@
 -(BOOL)isCaptain;
 -(BOOL)isAdmiral;
 -(BOOL)isFleetCaptain;
+-(BOOL)isOfficer;
 -(BOOL)isPlaceholder;
 -(BOOL)isUnique;
 -(BOOL)isDominion;
@@ -37,6 +38,7 @@
 -(NSString*)upSortType;
 -(NSString*)typeCode;
 -(NSString*)plainDescription;
+-(NSString*)disambiguatedTitle;
 -(int)costForShip:(DockEquippedShip*)equippedShip;
 -(int)costForShip:(DockEquippedShip*)equippedShip equippedUpgade:(DockEquippedUpgrade*)equippedUpgrade;
 @end