+++ src/DockUpgrade+Addons.m
@@ -112,6 +112,28 @@
     return [NSString stringWithFormat: @"%@ (%@)", self.title, self.upType];
 }
 
+-(NSString*)disambiguatedTitle
+{
+    NSString* externalId = self.externalId;
+
+    if ([externalId isEqualToString: @"quark_71786"]) {
+        return [NSString stringWithFormat: @"%@ (Tech)", self.title];
+    }
+    if ([externalId isEqualToString: @"quark_weapon_71786"]) {
+        return [NSString stringWithFormat: @"%@ (Weapon)", self.title];
+    }
+    if ([externalId isEqualToString: @"vulcan_high_command_2_0_71446"]) {
+        return [NSString stringWithFormat: @"%@ (2/0)", self.title];
+    }
+    if ([externalId isEqualToString: @"vulcan_high_command_1_1_71446"]) {
+        return [NSString stringWithFormat: @"%@ (1/1)", self.title];
+    }
+    if ([externalId isEqualToString: @"vulcan_high_command_0_2_71446"]) {
+        return [NSString stringWithFormat: @"%@ (0/2)", self.title];
+    }
+    return self.title;
+}
+
 -(BOOL)isTalent
 {
     return [self.upType isEqualToString: @"Talent"];
@@ -140,6 +162,11 @@
 -(BOOL)isFleetCaptain
 {
     return [self.upType isEqualToString: kFleetCaptainUpgradeType];
+}
+
+-(BOOL)isOfficer
+{
+    return [self.upType isEqualToString: kOfficerUpgradeType];
 }
 
 -(BOOL)isTech
@@ -202,6 +229,11 @@
     return YES;
 }
 
+-(BOOL)isIndependent
+{
+    return [self.faction isEqualToString: @"Independent"];
+}
+
 -(NSComparisonResult)compareTo:(DockUpgrade*)other
 {
     NSString* upTypeMe = [self upSortType];
@@ -246,6 +278,10 @@
 
     if ([self isFleetCaptain]) {
         return [targetShip fleetCaptainCount];
+    }
+
+    if ([self isOfficer]) {
+        return [targetShip officerLimit];
     }
 
     if ([self isTalent]) {
@@ -311,7 +347,7 @@
     }
 
     if ([self isCaptain]) {
-        return @"AAACaptain";
+        return @"AAAAAAAACaptain";
     }
 
     if ([self isTalent]) {
@@ -361,6 +397,10 @@
         return @"B";
     }
 
+    if ([self isOfficer]) {
+        return @"O";
+    }
+
     return @"?";
 
 }
@@ -399,6 +439,7 @@
 -(int)costForShip:(DockEquippedShip*)equippedShip equippedUpgade:(DockEquippedUpgrade*)equippedUpgrade
 {
     DockUpgrade* upgrade = self;
+    NSString* externalId = self.externalId;
 
     NSString* fleetCaptainSpecial = [[equippedShip.squad.equippedFleetCaptain upgrade] special];
 
@@ -572,7 +613,7 @@
         }
     }
 
-    if (!factionsMatch(ship, self) && !equippedShip.isResourceSideboard && !factionsMatch(self, equippedShip.flagship)) {
+    if (![upgrade isOfficer] && !factionsMatch(ship, self) && !equippedShip.isResourceSideboard && !factionsMatch(self, equippedShip.flagship)) {
         if ([captainSpecial isEqualToString: @"UpgradesIgnoreFactionPenalty"] && ![upgrade isCaptain] && ![upgrade isAdmiral]) {
             // do nothing
         } else if ([captainSpecial isEqualToString: @"NoPenaltyOnFederationOrBajoranShip"]  && [upgrade isCaptain]) {
@@ -589,6 +630,8 @@
                    [upgrade isFactionBorg]) {
         } else if ([captainSpecial isEqualToString: @"lore_71522"] &&
                    [upgrade isTalent]) {
+        } else if ([externalId isEqualToString: @"elim_garak_71786"]) {
+        } else if ([fleetCaptainOnThisShip isIndependent] && [ship isIndependent] && [upgrade isCaptain]) {
         } else {
             if (upgrade.isAdmiral) {
                 cost += 3;
@@ -622,14 +665,18 @@
 -(int)additionalWeaponSlots
 {
     NSString* special = self.special;
+    NSString* externalId = self.externalId;
 
     if ([special isEqualToString: @"AddTwoWeaponSlots"]) {
         return 2;
     }
-    if ([special isEqualToString: @"AddsOneWeaponOneTech"]) {
+    if ([special isEqualToString: @"AddsOneWeaponOneTech"] || [special isEqualToString: @"addoneweaponslot"]) {
         return 1;
     }
     if ([special isEqualToString: @"sakonna_gavroche"]) {
+        return 1;
+    }
+    if ([externalId isEqualToString: @"quark_weapon_71786"]) {
         return 1;
     }
     return 0;
@@ -637,13 +684,19 @@
 
 -(int)additionalTechSlots
 {
-    if ([self.special isEqualToString: @"AddsOneWeaponOneTech"]) {
+    NSString* special = self.special;
+    NSString* externalId = self.externalId;
+
+    if ([special isEqualToString: @"AddsOneWeaponOneTech"]) {
         return 1;
     }
-    if ([self.externalId isEqualToString: @"vulcan_high_command_2_0_71446"]) {
+    if ([externalId isEqualToString: @"vulcan_high_command_2_0_71446"]) {
         return 2;
     }
-    if ([self.externalId isEqualToString: @"vulcan_high_command_1_1_71446"]) {
+    if ([special isEqualToString: @"addonetechslot"] || [externalId isEqualToString: @"vulcan_high_command_1_1_71446"]) {
+        return 1;
+    }
+    if ([externalId isEqualToString: @"quark_71786"]) {
         return 1;
     }
     return 0;
@@ -664,6 +717,9 @@
 
 -(int)additionalTalentSlots
 {
+    if ([self.externalId isEqualToString: @"elim_garak_71786"]) {
+        return 1;
+    }
     return 0;
 }